requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.1.1
python-dotenv==1.0.0
streamlit>=1.26.0
streamlit-chat>=0.0.2.2
